Move the homeassistant sensor server under senser/homeassistant. Preparation for some kind of blog post
