+++ src/main/java/com/pi4j/examples/sensor/homeassistant/README.md
